+++ test/smoke/testApps/CoreAndFilter/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/CoreAndFilterTests.java
@@ -255,8 +255,7 @@ public class CoreAndFilterTests extends AiSmokeTest {
         List<Envelope> rdList = mockedIngestion.waitForItems("RequestData", 1);
 
         Envelope rdEnvelope = rdList.get(0);
-        String operationId = rdEnvelope.getTags().get("ai.operation.id");
-        List<Envelope> pvdList = mockedIngestion.waitForItemsInOperation("PageViewData", 3, operationId);
+        List<Envelope> pvdList = mockedIngestion.waitForItems("PageViewData", 3);
         assertEquals(0, mockedIngestion.getCountForType("EventData"));
 
         RequestData rd = (RequestData) ((Data<?>) rdEnvelope.getData()).getBaseData();
@@ -292,6 +291,7 @@ public class CoreAndFilterTests extends AiSmokeTest {
 
         assertNotNull(pv2);
         assertEquals(new Duration(123456), pv2.getDuration());
+        assertEquals("2010-10-10T00:00:00.000+0000", pvdEnvelope2.getTime());
         assertEquals("value", pv2.getProperties().get("key"));
         assertEquals("a-value", pv2.getProperties().get("a-prop"));
         assertEquals("another-value", pv2.getProperties().get("another-prop"));
@@ -312,6 +312,7 @@ public class CoreAndFilterTests extends AiSmokeTest {
 
         assertNotNull(pv3);
         assertEquals(new Duration(123456), pv3.getDuration());
+        assertEquals("2010-10-10T00:00:00.000+0000", pvdEnvelope3.getTime());
         assertEquals("value", pv3.getProperties().get("key"));
         assertEquals("a-value", pv3.getProperties().get("a-prop"));
         assertEquals("another-value", pv3.getProperties().get("another-prop"));
@@ -330,8 +331,14 @@ public class CoreAndFilterTests extends AiSmokeTest {
         assertTrue(pvdEnvelope3.getTags().get("ai.internal.sdkVersion").startsWith("java:3."));
 
         assertParentChild(rd, rdEnvelope, pvdEnvelope1, "GET /CoreAndFilter/trackPageView");
-        assertParentChild(rd, rdEnvelope, pvdEnvelope2, "GET /CoreAndFilter/trackPageView", "operation-name-goes-here");
-        assertParentChild(rd, rdEnvelope, pvdEnvelope3, "GET /CoreAndFilter/trackPageView", "operation-name-goes-here");
+
+        assertEquals("operation-id-goes-here", pvdEnvelope2.getTags().get("ai.operation.id"));
+        assertEquals("operation-parent-id-goes-here", pvdEnvelope2.getTags().get("ai.operation.parentId"));
+        assertEquals("operation-name-goes-here", pvdEnvelope2.getTags().get("ai.operation.name"));
+
+        assertEquals("operation-id-goes-here", pvdEnvelope3.getTags().get("ai.operation.id"));
+        assertEquals("operation-parent-id-goes-here", pvdEnvelope3.getTags().get("ai.operation.parentId"));
+        assertEquals("operation-name-goes-here", pvdEnvelope3.getTags().get("ai.operation.name"));
     }
 
     @Test