+++ test/smoke/testApps/SpringBootAttachInMain/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/SpringBootAttachInMainTest.java
@@ -1,0 +1,28 @@
+package com.microsoft.applicationinsights.smoketest;
+
+import java.util.List;
+
+import com.microsoft.applicationinsights.internal.schemav2.Data;
+import com.microsoft.applicationinsights.internal.schemav2.Envelope;
+import com.microsoft.applicationinsights.internal.schemav2.RequestData;
+import org.junit.*;
+
+import static org.junit.Assert.*;
+
+public class SpringBootAttachInMainTest extends AiSmokeTest {
+
+    @Test
+    @TargetUri("/test")
+    public void doMostBasicTest() throws Exception {
+        List<Envelope> rdList = mockedIngestion.waitForItems("RequestData", 1);
+
+        Envelope rdEnvelope = rdList.get(0);
+
+        RequestData rd = (RequestData) ((Data<?>) rdEnvelope.getData()).getBaseData();
+
+        assertEquals("GET /test", rd.getName());
+        assertEquals("200", rd.getResponseCode());
+        assertTrue(rd.getProperties().isEmpty());
+        assertTrue(rd.getSuccess());
+    }
+}