+++ test/smoke/testApps/TelemetryProcessors/src/main/java/com/microsoft/applicationinsights/smoketestapp/SpringBootApp.java
@@ -6,7 +6,6 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
 
 @SpringBootApplication
 public class SpringBootApp extends SpringBootServletInitializer {
-
     @Override
     protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
         return applicationBuilder.sources(SpringBootApp.class);